show port on "invoker list" command
This is to show port number of a process(if available) when "invoker
list" command is fired. This could be helpful when we do not mention a
specific port in .ini file, let invoker handle it and then need to
access the port, e.g access the process in "localhost:port" domain.